Make the diff more readable and change the diff direction
For me it's confusing that we are not showing what is the diff to be applied.
I made some coloring changes:
 - red values would be deleted or show old values
 - green values are new values
 - yellow shows a key is changed

The keys won't be duplicated anymore so it's easier to see what changes.
Also values changes are easier to see with the old => new format (red and green color respectively).