De-duplicate dynamic metrics
Some kube manifests define multiple of the same resource. This is
invalid, but shouldn't break kontrast.

Where kontrast meets duplicates, it tries to create the same Gauge
multiple times with the same label set. This results in this error when scraping metrics:

```
An error has occurred during metrics gathering:
2 error(s) occurred:
* collected metric kontrast_current_diffs label:<name:"object" value:"CustomResourceDefinition/verticalpodautoscalers.autoscaling.k8s.io" > label:<name:"object_ns" value:"default" > gauge:<value:1 >  was collected before with the same name and label values
* collected metric kontrast_current_diffs label:<name:"object" value:"CustomResourceDefinition/verticalpodautoscalercheckpoints.autoscaling.k8s.io" > label:<name:"object_ns" value:"default" > gauge:<value:1 >  was collected before with the same name and label values
```

To fix this, we de-duplicate the diffed resources first in a map
before creating metrics.